Python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
click==8.1.7
    # via uvicorn
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
fastapi==0.112.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
importlib-metadata==8.0.0
    # via opentelemetry-api
joblib==1.4.2
    # via
    #   -r requirements.in
    #   scikit-learn
numpy==1.26.4
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
opentelemetry-api==1.26.0
    # via
    #   -r requirements.in
    #   opentelemetry-distro
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-distro==0.47b0
    # via -r requirements.in
opentelemetry-instrumentation==0.47b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.47b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.47b0
    # via -r requirements.in
opentelemetry-sdk==1.26.0
    # via opentelemetry-distro
opentelemetry-semantic-conventions==0.47b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.47b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
pandas==2.2.2
    # via -r requirements.in
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-multipart==0.0.9
    # via -r requirements.in
pytz==2024.1
    # via pandas
scikit-learn==1.5.1
    # via -r requirements.in
scipy==1.14.0
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.38.2
    # via fastapi
threadpoolctl==3.5.0
    # via scikit-learn
typing-extensions==4.12.2
    # via
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
uvicorn==0.30.6
    # via -r requirements.in
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.20.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
